requirements.txt
pytest
pylint
ipdb
jupyterlab
numpy
pandas
matplotlib
seaborn
scikit-learn
jupyter
python-dotenv
darts
requests_cache
google-search-results
pytest
uvicorn
fastapi
requests
mlflow==2.1.1
prefect==2.14.9
google-cloud-bigquery
google-cloud-storage

#tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
#tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
#tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips
